requirements.txt
torch
torchvision
flask
Pillow
numpy
diffusers
transformers
compel
opencv-python
requests
accelerate
matplotlib
